requirements.txt
typer==0.9.0
Pillow==9.5.0
Cython==0.29.35
pytest==7.3.2